requirements.txt
annotated-types==0.6.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cssselect2==0.7.0
Django==5.0.2
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap-form==3.4
django-bootstrap3==23.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-jquery==3.1.0
django-jquery-ui==1.11.4.1
django-multiupload==0.6.1
django-otp==1.3.0
django-paypal==2.0
et-xmlfile==1.1.0
html5lib==1.1
idna==3.6
lxml==5.1.0
openpyxl==3.1.2
oscrypto==1.3.0
paypalrestsdk==1.13.3
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pyHanko==0.23.0
pyhanko-certvalidator==0.26.3
pyOpenSSL==24.0.0
pypdf==4.1.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.9.0
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
webencodings==0.5.1
xhtml2pdf==0.2.15
